+++ app/(landing-page)/CaseStudy.tsx
