+++ app/(landing-page)/Hero.module.scss
