+++ app/(landing-page)/MobileCards/HeroCardsMobileFormat.jsx
