+++ app/our-team/page.tsx
@@ -1,23 +1,22 @@
 import "../globals.css";
+
 import Footer from "@/components/Footer";
-import Navbar from "@/components/Navbar";
-import Contact from "@/sections/our-team/Contact";
-import Designers from "@/sections/our-team/Designers";
-import Developers from "@/sections/our-team/Developers";
-import Hero from "@/sections/our-team/Hero";
-import OurStory from "@/sections/our-team/OurStory";
-import OurStory2 from "@/sections/our-team/OurStory(dev)";
+import Contact from "./sections/Contact";
+import Designers from "./sections/Designers";
+import Developers from "./sections/Developers";
+import Hero from "./sections/Hero";
+import OurStory from "./sections/OurStory";
 import { Metadata } from "next";
 
 export const metadata: Metadata = {
   title: "Team",
 };
 
-const page = () => {
+const Page = () => {
   return (
     <div>
       <Hero />
-      <OurStory2 />
+      <OurStory />
       <Designers />
       <section className="">
         <Developers />
@@ -31,4 +30,4 @@ const page = () => {
   );
 };
 
-export default page;
+export default Page;
