+++ app/our-team/sections/OurStory.tsx
@@ -1,8 +1,9 @@
 "use client";
 
 import Image from "next/image";
-import styles from "../../components/scss/OurStory.module.scss";
 import Button from "@/components/Button";
+
+import styles from "../../../components/scss/OurStory.module.scss";
 
 const OurStory2 = () => {
   return (
